+++ src/tests/specs/orbiter/orbiter.http.page-views.spec.ts
@@ -6,7 +6,6 @@ import type {
 import { idlFactory as idlFactorOrbiter } from '$declarations/orbiter/orbiter.factory.did';
 import type { HttpRequest } from '$declarations/satellite/satellite.did';
 import { Ed25519KeyIdentity } from '@dfinity/identity';
-import type { Principal } from '@dfinity/principal';
 import { fromNullable, jsonReviver, toNullable } from '@dfinity/utils';
 import { type Actor, PocketIc } from '@hadronous/pic';
 import { nanoid } from 'nanoid';
@@ -17,14 +16,12 @@ import {
 	satelliteIdMock,
 	type SetPageViewPayload
 } from '../../mocks/orbiter.mocks';
-import { assertCertification } from '../../utils/certification-test.utils';
 import { toBodyJson } from '../../utils/orbiter-test.utils';
 import { tick } from '../../utils/pic-tests.utils';
 import { controllersInitArgs, ORBITER_WASM_PATH } from '../../utils/setup-tests.utils';
 
-describe('Orbiter > Http', () => {
+describe('Orbiter > HTTP > Page views', () => {
 	let pic: PocketIc;
-	let canisterId: Principal;
 	let actor: Actor<OrbiterActor>;
 
 	const controller = Ed25519KeyIdentity.generate();
@@ -34,8 +31,6 @@ describe('Orbiter > Http', () => {
 	const NON_POST_METHODS = ['GET', 'PUT', 'PATCH', 'DELETE'];
 
 	const RESPONSE_OK = { ok: { data: null } };
-	const RESPONSE_404 = { err: { code: 404, message: 'Not found' } };
-	const RESPONSE_405 = { err: { code: 405, message: 'Method not allowed' } };
 	const RESPONSE_500_NO_VERSION_PROVIDED = {
 		err: { code: 500, message: 'juno.error.no_version_provided' }
 	};
@@ -53,7 +48,6 @@ describe('Orbiter > Http', () => {
 		});
 
 		actor = c;
-		canisterId = cId;
 
 		// Certified responses are initialized asynchronously
 		await tick(pic);
@@ -63,118 +57,7 @@ describe('Orbiter > Http', () => {
 		await pic?.tearDown();
 	});
 
-	describe.each(['/something', '/'])('Route not found for %s', (url) => {
-		it.each([...NON_POST_METHODS, 'POST'])(
-			'should return a certified not found response for %s',
-			async (method) => {
-				const { http_request } = actor;
-
-				const request: HttpRequest = {
-					body: [],
-					certificate_version: toNullable(2),
-					headers: [],
-					method,
-					url
-				};
-
-				const response = await http_request(request);
-
-				expect(
-					response.headers.find(([key, _value]) => key.toLowerCase() === 'content-type')?.[1]
-				).toEqual('application/json');
-
-				const decoder = new TextDecoder();
-				const responseBody = decoder.decode(response.body as Uint8Array<ArrayBufferLike>);
-
-				expect(JSON.parse(responseBody)).toEqual(RESPONSE_404);
-
-				await assertCertification({
-					canisterId,
-					pic,
-					request,
-					response,
-					currentDate,
-					statusCode: 404
-				});
-			}
-		);
-	});
-
-	describe.each(['/view/something'])('Sub-route not found for %s', (url) => {
-		it.each(NON_POST_METHODS)(
-			'should return a certified not found response for %s',
-			async (method) => {
-				const { http_request } = actor;
-
-				const request: HttpRequest = {
-					body: [],
-					certificate_version: toNullable(2),
-					headers: [],
-					method,
-					url
-				};
-
-				const response = await http_request(request);
-
-				expect(
-					response.headers.find(([key, _value]) => key.toLowerCase() === 'content-type')?.[1]
-				).toEqual('application/json');
-
-				const decoder = new TextDecoder();
-				const responseBody = decoder.decode(response.body as Uint8Array<ArrayBufferLike>);
-
-				expect(JSON.parse(responseBody)).toEqual(RESPONSE_404);
-
-				await assertCertification({
-					canisterId,
-					pic,
-					request,
-					response,
-					currentDate,
-					statusCode: 404
-				});
-			}
-		);
-	});
-
-	describe.each(['/view', '/views'])('Route not allowed for %s', (url) => {
-		it.each(NON_POST_METHODS)(
-			'should return a certified not allowed response for %s',
-			async (method) => {
-				const { http_request } = actor;
-
-				const request: HttpRequest = {
-					body: [],
-					certificate_version: toNullable(2),
-					headers: [],
-					method,
-					url
-				};
-
-				const response = await http_request(request);
-
-				expect(
-					response.headers.find(([key, _value]) => key.toLowerCase() === 'content-type')?.[1]
-				).toEqual('application/json');
-
-				const decoder = new TextDecoder();
-				const responseBody = decoder.decode(response.body as Uint8Array<ArrayBufferLike>);
-
-				expect(JSON.parse(responseBody)).toEqual(RESPONSE_405);
-
-				await assertCertification({
-					canisterId,
-					pic,
-					request,
-					response,
-					currentDate,
-					statusCode: 405
-				});
-			}
-		);
-	});
-
-	describe('configured', () => {
+	describe('With configuration', () => {
 		interface SetPageViewRequest {
 			key: AnalyticKey;
 			page_view: SetPageViewPayload;
